Converted long counters to double so the decimals work correctly

+++ lib/src/main/kotlin/nl/mirrajabi/humanize/duration/languages/EnglishDictionary.kt
@@ -13,7 +13,7 @@ class EnglishDictionary : LanguageDictionary {
         KeyConstants.DECIMAL to "."
     )
 
-    override fun get(key: String, count: Long): String {
-        return words[key] + (if (count != 1L && key != KeyConstants.DECIMAL) "s" else "")
+    override fun get(key: String, count: Double): String {
+        return words[key] + (if (count != 1.0 && key != KeyConstants.DECIMAL) "s" else "")
     }
 }
